led-indicator: hcgw: Control RF LEDs by Linux
Unlike the old RM FW stack, the one based on Zephyr does no longer decide on
its own when to turn on the RF LEDs. Instead, it just relays the lines coming
from Linux.

This aligns the behavior of the HCGW with the one found on the LCGW, which is
preferable from a maintenance perspective.

Issue: SG-20159