sshtunnel: Add sshtunnel-timer package for factory image
When the customer powers up a gateway that has the factory image
installed, it should immediately upgrade to the regular image and
therefore, only run for a very short time. If this is not the cases,
this package will make sure the sshtunnel gets started. Maintenance
access will then be the only way to interact with the gateway as the
factory image does not contain the cloudadapter service.

Issue: SG-20494