manufacturing-tools: Fix LED behavior
Without this change, a PCBA that has not yet passed the FCT but is successfully
bootstrapped, will not turn on all three green LEDs after a reboot. This
created quite some confusion at the assembly line, namely during shift changes,
and put avoidable load on the service department.

Issue: SG-18829