coap-transport-proxy: Update to version 0.6.1
Migrate to `cargo-update-recipe-crates`

Use the tools provided by Yocto instead of `cargo-bitbake`.
This eliminates the need for `BB_STRICT_CHECKSUM = "0"` in
recipes for Rust applications.
This change also removes the dependency to `cargo-bitbake` that
is not maintained very actively.

Also inherit from pkgconfig, otherwise the build will fail due to
pkg-config not being available.

Add THIRDPARTY.toml that was created with the cargo bundle-licenses
command.

Issue: SG-18399