gateway-config-backend: Adapt to Yocto Nanbield
This commit contains necessary changes for the Yocto Nanbield release.

Also inherit from pkgconfig, otherwise the build will fail due to
pkg-config not being available.

Migrate to `cargo-update-recipe-crates` to use the tools provided by
Yocto instead of `cargo-bitbake`.  This eliminates the need for
`BB_STRICT_CHECKSUM = "0"` in recipes for Rust applications.  This
change also removes the dependency to `cargo-bitbake` that is not
maintained very actively.

Issue: SG-18399