python3-opentelemetry-api: Patch bad importlib import
Since version 1.17.0, opentelemetry-api uses a pinned version of
imporlib-metdata independend of the Python version. As a result, if there
is a compatible version of importlib-metadata available in the std
Python lib, the package would still require the backported package.

The upstream recipe of importlib-metadata pulls in a lot of
dependencies. As a result this hack will just patch opentelemetry-api to
use metadata from the std lib.

Issue: SG-21173