swupdate-legacy: Add recipe and class for SWUpdate 2020.11
For the new Gateway software rollout concept, `swupdate` shall be
updated and adapted to be able to connect to an update server. To keep
the old way of updating for "legacy" images, we need a recipe with a
different name.

The `swupdate` class has some common dependencies (`swupdate` library).
Therefore, it is copied as well.

The files were copied (more or less as-is) from commit 0c84450 ("Revert
"Drop support for SWUpdate 2020.11"") in the
smart-garden-gateway-yocto-meta-swupdate repository.

Issue: SG-21568