healthcheck: make ppp0-related tests depend on ppp0 interface
By only executing tests that depend on ppp0 if the ppp0 interface is
available, we get sensible results for the prevalence of the actual
issues being tested. The test test_ppp0_interface_up is executed
unconditionally to find gateways with fundamental issues with the
radio module. This can e.g. be a defective radio module hardware, or
an incorrectly flashed firmware. In the past, such an issue caused all
ppp0-related tests to fail as well, thus tainting the results.