consent: Collect support data by default
In the future, the customer can opt-in to allow us to collect logs, metrics
and support bundles (snapshots) on the gateway. This will be reflected
by the presence of the files `consent-to-telemetry` and
`consent-to-support-bundle` in `/etc`.
We install these files now to keep the current behaviour. But they will
be removed from the image in a future release. Then the customer needs
to opt-in and the files will be written through the back-end.
We keep the files written by the customer between updates.

Issue: SG-22111