cloudadapter: Set device/0/battery_level limit
Reported by different mowers, limit is mostly breached by CBT11 (2474 devices)
as it reports battery_level in 5% steps. This is followed by CBTG (855 devices).
Other mower very rarely breach the limit.

In 90% of the cases the resource is reported as a single resource update,
in another 9% it is only grouped with battery_status.

Saves 100'000 events per day (that mostly update a very expensive shadow).

Issue: SG-22905